feat: add usechangeurl hook
