+++ src/components/GraphiQLClient/VariablesEditor.tsx
@@ -1,24 +1,15 @@
 import { useFormContext, Controller } from "react-hook-form";
-import { encodeBase64 } from "../../utils/encodeBase64";
-import { useRouter } from "next/router";
+import { useChangeURL } from "../../hooks/changeURL";
 
 const VariablesEditor: React.FC = () => {
   const { control, getValues } = useFormContext();
+  const { changeURL } = useChangeURL();
   const DEFAULT_VARIABLES = `{"id": 1}`;
 
-  const router = useRouter();
-
   const updateUrl = (variables) => {
-    const queryParams = router.asPath.split("?")[1];
-    const endpoint = encodeBase64(getValues("url"));
+    const endpoint = getValues("url");
     const query = getValues("query");
-    const body = encodeBase64(
-      JSON.stringify({ query, ...(variables && { variables }) }),
-    );
-    const newURL = queryParams
-      ? `/graphql/${endpoint}/${body}?${queryParams}`
-      : `/graphql/${endpoint}/${body}`;
-    router.replace({}, newURL, { shallow: true });
+    changeURL({ variables, url: endpoint, query });
   };
 
   return (
